feat: test for input and output flag
