+++ app/index.js
@@ -1,5 +1,8 @@
 const { processes } = require("./process");
-const { getConfig, getInput, getOutput } = require("./utils/utils");
+const getConfig = require("./utils/getConfig.js");
+const getInput = require("./utils/getInput.js");
+const getOutput = require("./utils/getOutput.js");
+
 
 //!_____________может принимать -c и -config добавить проверку______
